feat: 상품 상세 페이지 리팩토링 및 상태 관리 개선
